snapvault/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
azure-cosmos==4.7.0
azure-storage-blob==12.20.0
requests==2.31.0
python-dotenv==1.0.1
